epubToWav/requirements.txt
epub2txt
python-dotenv
requests
